requirements.txt
azure-ai-ml==1.18.0
azure-identity==1.17.1
azureml-mlflow==1.56.0
boto3
jwt==1.3.1
mlflow==2.14.2
psycopg2-binary
python-dotenv==1.0.1
